pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sb3-mujoco-suite"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "stable-baselines3>=2.3.0",
    "gymnasium[mujoco]>=0.29.1",
    "mujoco>=3.0.0",
    "torch",
    "numpy",
    "tensorboard",
    "imageio",
    "pyyaml>=6.0.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
train-ant = "rl_core.train:main"

